# docs/swagger.yaml
openapi: 3.0.1
info:
  title: Open API
  version: v1
servers:
  - url: http://localhost:5000
paths:
  /api/v1/choice-templates:
    get:
      tags:
        - ChoiceTemplate
      summary: 点検種別の一覧を取得する
      responses:
        '200':
          description: 取得に成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChoiceTemplate'
        '500':
          description: サーバー内部エラー
    post:
      tags:
        - ChoiceTemplate
      summary: 点検種別を作成する
      requestBody:
        description: 作成用点検種別データ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoiceTemplate'
        required: true
      responses:
        '201':
          description: 作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoiceTemplate'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
  '/api/v1/choice-templates/{id}':
    get:
      tags:
        - ChoiceTemplate
      summary: 指定の点検種別を取得する
      parameters:
        - name: id
          in: path
          description: 指定の点検種別に紐づくID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoiceTemplate'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象リソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
    put:
      tags:
        - ChoiceTemplate
      summary: 指定の点検種別を更新する
      parameters:
        - name: id
          in: path
          description: 指定の点検種別に紐づくID
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: 更新用点検種別データ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChoiceTemplate'
        required: true
      responses:
        '201':
          description: 更新に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoiceTemplate'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象リソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
    delete:
      tags:
        - ChoiceTemplate
      summary: 指定の点検種別を削除する
      parameters:
        - name: id
          in: path
          description: 指定の点検種別に紐づくID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 削除に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChoiceTemplate'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象リソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
...
  /api/v1/inspection-groups:
    get:
      tags:
        - InspectionGroup
      summary: 点検グループの一覧を取得する
      responses:
        '200':
          description: 取得に成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InspectionGroup'
        '500':
          description: サーバー内部エラー
    post:
      tags:
        - InspectionGroup
      summary: 点検グループを作成する
      requestBody:
        description: 作成用点検グループデータ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionGroup'
        required: true
      responses:
        '201':
          description: 作成に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionGroup'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
  '/api/v1/inspection-groups/{id}':
    get:
      tags:
        - InspectionGroup
      summary: 指定の点検グループを取得する
      parameters:
        - name: id
          in: path
          description: 指定の点検グループに紐づくID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 取得に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionGroup'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象リソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
    put:
      tags:
        - InspectionGroup
      summary: 指定の点検グループを更新する
      parameters:
        - name: id
          in: path
          description: 指定の点検グループに紐づくID
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: 更新用点検グループデータ
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionGroup'
        required: true
      responses:
        '201':
          description: 更新に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionGroup'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象リソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
    delete:
      tags:
        - InspectionGroup
      summary: 指定の点検グループを削除する
      parameters:
        - name: id
          in: path
          description: 指定の点検グループに紐づくID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: 削除に成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionGroup'
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: 対象リソースが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: サーバー内部エラー
...
components:
  schemas:
...
    ChoiceTemplate:
      required:
        - choices
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        choices:
          type: array
          items:
            $ref: '#/components/schemas/Option'
      additionalProperties: false
...
    InspectionGroup:
      required:
        - description
        - id
      type: object
      properties:
        id:
          type: integer
          format: int32
        description:
          type: string
      additionalProperties: false
...
    Option:
      required:
        - description
        - option_id
        - order_index
      type: object
      properties:
        option_id:
          type: integer
          format: int32
        order_index:
          type: integer
          format: int32
        description:
          type: string
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }