# src/ClientApp/src/typescript-fetch/models/ChoiceTemplate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Open API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Option,
    OptionFromJSON,
    OptionFromJSONTyped,
    OptionToJSON,
} from './Option';

/**
 * 
 * @export
 * @interface ChoiceTemplate
 */
export interface ChoiceTemplate {
    /**
     * 
     * @type {number}
     * @memberof ChoiceTemplate
     */
    id: number;
    /**
     * 
     * @type {Array<Option>}
     * @memberof ChoiceTemplate
     */
    choices: Array<Option>;
}

export function ChoiceTemplateFromJSON(json: any): ChoiceTemplate {
    return ChoiceTemplateFromJSONTyped(json, false);
}

export function ChoiceTemplateFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChoiceTemplate {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'choices': ((json['choices'] as Array<any>).map(OptionFromJSON)),
    };
}

export function ChoiceTemplateToJSON(value?: ChoiceTemplate | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'choices': ((value.choices as Array<any>).map(OptionToJSON)),
    };
}