# src/ClientApp/src/typescript-fetch/models/InspectionGroup.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Open API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface InspectionGroup
 */
export interface InspectionGroup {
    /**
     * 
     * @type {number}
     * @memberof InspectionGroup
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof InspectionGroup
     */
    description: string;
}

export function InspectionGroupFromJSON(json: any): InspectionGroup {
    return InspectionGroupFromJSONTyped(json, false);
}

export function InspectionGroupFromJSONTyped(json: any, ignoreDiscriminator: boolean): InspectionGroup {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'description': json['description'],
    };
}

export function InspectionGroupToJSON(value?: InspectionGroup | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'description': value.description,
    };
}